Fix for DonationScreenModelTest

+++ shared/src/commonTest/kotlin/mobi/cwiklinski/bloodline/ui/model/DonationScreenModelTest.kt
@@ -10,11 +10,11 @@ import kotlinx.coroutines.test.TestCoroutineScheduler
 import kotlinx.coroutines.test.UnconfinedTestDispatcher
 import kotlinx.coroutines.test.runTest
 import kotlinx.coroutines.test.setMain
+import mobi.cwiklinski.bloodline.common.manager.AppCallbackManager
 import mobi.cwiklinski.bloodline.data.filed.CenterServiceImplementation
 import mobi.cwiklinski.bloodline.data.filed.DonationServiceImplementation
 import mobi.cwiklinski.bloodline.data.filed.DummyData
 import mobi.cwiklinski.bloodline.data.filed.ProfileServiceImplementation
-import mobi.cwiklinski.bloodline.common.manager.AppCallbackManager
 import mobi.cwiklinski.bloodline.ui.util.UiTestTools
 import kotlin.test.BeforeTest
 import kotlin.test.Test
@@ -76,7 +76,7 @@ class DonationScreenModelTest {
             newDonation.date,
             newDonation.center,
             newDonation.type.type,
-            newDonation.hemoglobin.toInt(),
+            newDonation.hemoglobin,
             newDonation.systolic,
             newDonation.diastolic,
             newDonation.disqualification
